fix implicit conversion again

+++ Cesium3DTiles/src/upsampleGltfForRasterOverlays.cpp
@@ -680,10 +680,10 @@ namespace Cesium3DTiles {
                     position = ellipsoid.cartographicToCartesian(cartographic);
                     position -= center;
 
-                    for (int32_t c = 0; c < 3; ++c) {
+                    for (uint32_t c = 0; c < 3; ++c) {
 						output.push_back(static_cast<float>(position[c]));
-						attribute.minimums[c] = glm::min(attribute.minimums[c], position[c]);
-						attribute.maximums[c] = glm::max(attribute.maximums[c], position[c]);
+						attribute.minimums[c] = glm::min(attribute.minimums[c], position[static_cast<int32_t>(c)]);
+						attribute.maximums[c] = glm::max(attribute.maximums[c], position[static_cast<int32_t>(c)]);
                     }
                 }
                 else {
